Accounting/Deployed/requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2
click==8.1.7
MarkupSafe==2.1.3
itsdangerous==2.1.2
blinker==1.6.2
pandas>=1.5.0
pyarrow>=10.0.0
openpyxl>=3.0.0
python-dateutil>=2.8.0
bcrypt>=4.0.0
flask-wtf>=1.2.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
Flask-Limiter>=3.5.0
boto3>=1.34.0
APScheduler>=3.10.0
Flask-Caching>=2.1.0
python-json-logger>=2.0.0
alembic>=1.13.0
sqlalchemy>=2.0.0